Подправил визуализацию
